Draw Parameters for equation j

draw_parameters_j simulates from the posterior of the model's parameters for each equation \(j\) separately. The posterior is simulated using a Metropolis-within-Gibbs sampling procedure.

Usage

draw_parameters_j(
  y_matrix,
  x_matrix,
  character_gamma_matrix,
  character_beta_matrix,
  jx
)

Arguments

y_matrix

A \((T \times n)\) matrix \(Y\), where \(T\) is the number of observations and \(n\) the number of equations, i.e. endogenous variables.

x_matrix

A \((T \times k)\) matrix \(X\) of observations on \(k\) exogenous variables.

character_gamma_matrix

A matrix \(\Gamma\) that holds the coefficients in character form for all equations. The dimensions of the matrix are \((T \times n)\), where \(T\) is the number of observations and \(n\) the number of equations.

character_beta_matrix

A matrix \(\beta\) that holds the coefficients in character form for all equations. The dimensions of the matrix are \((k \times n)\), where \(k\) is the number of exogenous variables and \(n\) the number of equations.

jx

The index of equation \(j\).

Value

A list containing matrices for the saved draws of parameters and additional diagnostic information.

Details

The sampler works as follows:

Initialize sampler

Conditional on \(\Omega^{(w-1)}\) and the data draw \(delta_{\gamma}\).

Conditional on \(delta_{\gamma}^{(w)}\) draw \(\Omega^{(w)}\).

Conditional on \(Omega^{(w)}\) and \(\delta_{\gamma}^{(w)}\) draw \(\delta_\beta^{(w)}\).

Go back to step 2.