Quarterly time series of key macro variables for a small open economy (Switzerland)

Quarterly time series of key macro variables for a small open economy (Switzerland)

Usage

small_open_economy

Format

A list of 12 ts objects (quarterly):

consumption

Private consumption \(C_t\).

investment

Gross fixed capital formation \(I_t\).

exports

Exports \(X_t\).

imports

Imports \(M_t\).

prices

Price level \(P_t\).

interest_rate

Short‐term domestic interest rate \(R_t\).

gdp

Equilibrium output \(GDP_t\).

domestic_demand

Domestic demand \(D_t\).

world_gdp

World GDP (exogenous) \(WGDP_t\).

interest_rate_germany

German short‐term rate (exog) \(R_{GE,t}\).

exchange_rate

Nominal exchange rate (exog) \(ER_t\).

oil_price

Oil price (exogenous) \(OP_t\).

Source

https://www.seco.admin.ch/, https://kof.ethz.ch/